Added Menu class for showLoopingMenuUntiExit
