+++ src/main/java/seedu/duke/Ui.java
@@ -192,69 +192,5 @@ public class Ui {
         printMap(sortedDeliveries);
     }
 
-    /**
-     * Method backbone for menu selection
-     * Parser is only called for commands that require argument parsing
-     * @param deliveryman is the currently loaded profile
-     */
-    public void showLoopingMenuUntilExit(Deliveryman deliveryman) {
-        Parser parser = new Parser();
-        Scanner sc = new Scanner(System.in);
-        String userInput, userCommand, userArguments;
-        int deliveryNumber;
-        do {
-            promptUserInput();
-            userInput = sc.nextLine();
-            if (userInput.isBlank()){
-                userInput = "INVALID";
-            }
-            userCommand = parser.parseCommand(userInput);
-            userArguments = parser.parseArguments(userInput);
-            if (!userArguments.isEmpty()){
-                switch (userCommand){
-                case "edit":
-                    String inputProfileData = parser.parseInput("edit", userArguments,deliveryman);
-                    deliveryman.updateProfile(inputProfileData);
-                    break;
-                case "view":
-                    processViewDelivery(userArguments, deliveryman, parser);
-                    break;
-                case "complete":
-                    processCompleteDelivery(userArguments, deliveryman, parser);
-                    break;
-                default:
-                    System.out.println("Please enter a valid command!");
-                }
-            } else{
-                switch (userCommand) {
-                case "help":
-                    showHelpMessage();
-                    break;
-                case "profile":
-                    showProfile(deliveryman);
-                    break;
-                case "start":
-                    DeliveryList.load();
-                    break;
-                case "list":
-                    showDeliveryList();
-                    break;
-                case "record":
-                    showRecords(deliveryman.getRecords());
-                    break;
-                case "route":
-                case "deliveryroute":
-                    processDeliveryRoute();
-                    break;
-                case "bye":
-                    showFarewellScreen();
-                    break;
-                default:
-                    System.out.println("Please enter a valid command!");
-                }
-            }
-        } while (!userCommand.equalsIgnoreCase("bye"));
-    }
-
 
 }
